Reference

Index

VectorSpaceDarkMatter.FCoeffs

VectorSpaceDarkMatter.GaussianF

VectorSpaceDarkMatter.ModelDMSM

VectorSpaceDarkMatter.PartialRate

VectorSpaceDarkMatter.ProjectedF

VectorSpaceDarkMatter.Tophat

VectorSpaceDarkMatter.Wavelet

VectorSpaceDarkMatter.f_uSph

VectorSpaceDarkMatter.ProjectF

VectorSpaceDarkMatter.f2_norm

VectorSpaceDarkMatter.partial_rate

VectorSpaceDarkMatter.rate

VectorSpaceDarkMatter.readFnlm

VectorSpaceDarkMatter.readK

VectorSpaceDarkMatter.update!

VectorSpaceDarkMatter.writeFnlm

VectorSpaceDarkMatter.writeK

Docstrings

VectorSpaceDarkMatter.Wavelet — Type

Wavelet(umax)

Spherical Haar wavelets. Contains the maximum value of u = $|\vec{u}|$ that the basis of will be evaluated over. If umax is omitted, it will be set to 1.0

VectorSpaceDarkMatter.Tophat — Type

Tophat(xi[, umax])

Tophat basis functions between each point in xi. The range will be scaled by umax, the maximum value of u = $|\vec{u}|$. If umax is omitted, it will be set to 1.0

VectorSpaceDarkMatter.f_uSph — Type

f_uSph(f::Function; z_even=false, phi_even=false, 
    phi_cyclic=1, phi_symmetric=false)

Struct that adds decorations to a function $f(u, \theta, \phi)$ that tell ProjectF various properties about the function that speed up integration.

z_even : (boolean) if f_uSph(x,y,z) = f_uSph(x,y,-z). Implies $\langle \ell m | f \rangle = 0$ if $(\ell+m)$ is odd

phi_even : (boolean) if f_uSph(u,theta,phi) = f_uSph(u,theta,-phi) Implies $\langle \ell m | f \rangle$ = 0 if $m$ is odd

phi_cyclic : (integer) if f_uSph(u,theta,phi) = f_uSph(u,theta,phi + 2*pi/n)

phi_symmetric : (boolean) if f_uSph(u,theta) is independent of phi

VectorSpaceDarkMatter.GaussianF — Type

GaussianF(c::Float64, uSph::Vector{Float64}, sigma::Float64)

Contains parameters for a gaussian function where uSph is the spherical vector indicating the center of the gaussian, uSph = (u, theta, phi). sigma is the dispersion (equal to $\sqrt{2} \sigma$ in a "normal" gaussian). c is the overall amplitude. u and sigma should have the same units.

An instance g of GaussianF is callable by g(uvec), which will evaluate the gaussian at uvec = [u, theta, phi].

VectorSpaceDarkMatter.f2_norm — Function

f2_norm(g::GaussianF)

L2 norm of a single Gaussian $\int d^3 u \: g^2(u)$

f2_norm(g::Vector{GaussianF})

L2 norm of a set of Gaussians $\int d^3 u \: (\sum_i g_i(u))^2$

f2_norm(f::Function, a, b)

Integral $\int d^3 u \: f^2(u)$ over the range a = [u_min, theta_min, phi_min] to b = [u_max, theta_max, phi_max].

f2_norm(pf::T) where T<:Union{ProjectedF, FCoeffs}

$\int d^3 u \: f^2(u)$ for a ProjectedF or FCoeffs is equal to $u_{\textrm{max}}^3 \sum_{n \ell m} f_{n \ell m}^2$

VectorSpaceDarkMatter.ProjectF — Function

ProjectF(f, nl_max::Tuple{Int,Int}, radial_basis::RadialBasis; 
         dict=false, use_measurements=false, integ_method=:cubature,
         integ_params=(;))

Evaluates $\langle f | n \ell m \rangle$ at each $(n,\ell,m)$ up to nl_max = (n_max, l_max) and returns the result as a ProjectedF.

f : Can be a Function, f_uSph, GaussianF, or Vector{GaussianF}. f_uSph is preferred if your function has any symmetries, and is not gaussian, as specifying those will greatly speed up evaluation.

radial_basis : Either a Wavelet or Tophat

dict : If true, returns an FCoeffs instead of a ProjectedF, which stores the coefficients as a dictionary instead.

use_measurements : If true, will give the results as a measurement with uncertainty given by the integration error.

integ_method : Either :cubature, :vegas, or :vegasmc

integ_params : keyword arguments to pass to the integrator. If :cubature, these are kwargs for hcubature. If :vegas or :vegasmc, these are kwargs for MCIntegration's integrate method.

VectorSpaceDarkMatter.ProjectedF — Type

ProjectedF{A, B}(fnlm::Matrix{A}, lm, radial_basis::B)

Stores the $\langle f | n \ell m \rangle$ coefficients and the radial basis that was used to calculate them. It is assumed that spherical harmonics were used for the angular parts. A is the element type of the matrix that stores fnlm coefficients (should be either Float64 or Measurement). B is the type of radial basis. lm stores a vector of (ell,m) for the fnlm matrix: fnlm is indexed as [n+1,i] corresponding to (ell,m) = lm[i]

VectorSpaceDarkMatter.FCoeffs — Type

FCoeffs{A, B}(fnlm::Dict{Tuple{Int64,Int64,Int64}, A},
    radial_basis::B)

Stores the $\langle f | n \ell m \rangle$ coefficients as a dictionary with (n,ell,m) => f_nlm and the radial basis that was used to calculate them. It is assumed that spherical harmonics were used for the angular parts. A is the element type of the dict that stores fnlm coefficients (should be either Float64 or Measurement). B is the type of radial basis.

VectorSpaceDarkMatter.update! — Function

update!(fc::FCoeffs, f, nlm::Tuple{Int,Int,Int}; kwargs...)

Evaluates a particular $(n,\ell,m)$ coefficient for the function f and the radial basis fc.radial_basis and stores the result in fc.fnlm[(n,ell,m)]. Will overwrite existing data. kwargs correspond to the kwargs for getFnlm

If called with a vector of nlm tuples, runs update! for each (n,ell,m)

VectorSpaceDarkMatter.writeFnlm — Function

writeFnlm(outfile, pf)

Writes the coefficients from pf (either a ProjectedF or an FCoeffs) to file outfile.

VectorSpaceDarkMatter.readFnlm — Function

readFnlm(infile[, radial_basis::RadialBasis]; dict=false, use_err=true)

Reads coefficients from infile. Can optionally manually define the basis via radial_basis argument. Optional arguments:

dict : If true, will return an FCoeffs. If false, returns ProjectedF.

use_err : Whether or not to load the uncertainties on the fnlm values.

VectorSpaceDarkMatter.rate — Function

rate(R, model::ModelDMSM, pfv::ProjectedF, pfq::ProjectedF; 
    ell_max=nothing, use_measurements=false)

Calculates the rate for a given model, projected $g_\chi$ (pfv), projected $f_s^2$ (pfq), and rotation R given as a quaternion or rotor (if you want other rotations, see Quaternionic.jl's conversion methods). R can also be a vector of quaternions or rotors.

The maximum $\ell$ that is used is the minimum of the specified ell_max and the maximum $\ell$ for each of pfv and pfq.

Using measurements here is currently slow.

rate(R, mcK::PartialRate)

Calculates the rate for a given partial rate matrix mcK and rotation R given as a quaternion or rotor (to use other rotations, see Quaternionic.jl's conversion methods). R can also be a vector of quaternions or rotors.

The maximum $\ell$ that is used is the minimum of the specified ell_max and the maximum $\ell$ for each of pfv and pfq.

Measurements are not supported here (yet).

VectorSpaceDarkMatter.ModelDMSM — Type

ModelDMSM(fdm_n, mX, mSM, deltaE)

Stores the relevant model parameters.

fdm_n : The power $n$ of the dark matter form factor $(\alpha m_e / q)^{n}$

mX : dark matter mass in eV

mSM : mass of target particle in eV

deltaE : discrete final state energy in eV

VectorSpaceDarkMatter.partial_rate — Function

partial_rate(model::ModelDMSM, pfv::ProjectedF, pfq::ProjectedF; 
             ell_max=nothing, use_measurements=true)

Calculates the partial rate matrix, stored as a flattened vector, for a given model, velocity distribution (pfv), and material form factor (pfq).

The maximum $\ell$ that is used is the minimum of the specified ell_max and the maximum $\ell$ for each of pfv and pfq.

VectorSpaceDarkMatter.PartialRate — Type

PartialRate(K::Vector{A}, model::ModelDMSM, v_basis::RadialBasis,
            q_basis::RadialBasis)

Stores the flattened partial rate matrix along with the relevant model and basis parameters.

VectorSpaceDarkMatter.writeK — Function

writeK(outfile, mcK::PartialRate)

Writes a partial rate matrix $\mathcal{K}$ stored in mcK to outfile.

VectorSpaceDarkMatter.readK — Function

readK(infile[, v_basis::RadialBasis, q_basis::RadialBasis]; use_err=true)

Reads partial rate matrix from infile. Can optionally manually define the bases via v_basis and q_basis arguments, otherwise will attempt to read from the header of the csv file. Note that automatic basis reading only works for Wavelet bases.

use_err : Whether or not to load the uncertainties.